Moved Requirements to Requirement.hs and made minor improvements to Body.hs (Changes related)
