+++ code/stable/glassbr/Website/GlassBR_SRS.html
@@ -3196,7 +3196,7 @@ Functional Requirements
 <ul class="hide-list-style">
 <li>
 <div id="FR:Input-Glass-Props">
-Input-Glass-Props: Input the quantities from <a href=#Table:R1ReqInputs>R1ReqInputs</a>, which define the glass dimensions, type of glass, tolerable probability of failure and the characteristics of the blast. Note: <em>a</em> and <em>b</em> will be input in terms of millimetres and will be converted to the equivalent value in metres.
+Input-Glass-Props: Input the quantities from <a href=#Table:R1ReqInputs>R1ReqInputs</a>, which define the glass dimensions, type of glass, tolerable probability of failure, and the characteristics of the blast. Note: <em>a</em> and <em>b</em> will be input in terms of millimetres and will be converted to the equivalent value in metres.
 </div>
 </li>
 </ul>
